ARM: dts: msm: Waipio Configuration to set thread priority
In LL use-case GPR callback glink-adsp_apps thread
was in runnable state at the time of issue.
It means GPR callback is running in a normal thread which
just has 120 priority which is not RT thread, So
causes pop noise issue.
It needs to set glink-adsp_apps thread priority to be RT.
This Configuration enables Client to set the configuration thread
to be in RT.

(cherry-pick I2b6512ad06353cbcae30706fa27326210661dc0c)
Change-Id: I09937db8f4d866be150603aa65286a1b8b718b24
Signed-off-by: Shaik Jabida <quic_sjabida@quicinc.com>